requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
playwright==1.40.0
streamlit==1.28.1
pandas==2.1.3
plotly==5.17.0 